Refactor docstring miner by language modules (#131)
