+++ src/specleft/discovery/miners/python/__init__.py
@@ -1,0 +1,8 @@
+# SPDX-License-Identifier: Apache-2.0
+# Copyright (c) 2026 SpecLeft Contributors
+
+"""Python-specific discovery miners."""
+
+from specleft.discovery.miners.python.docstrings import extract_python_items
+
+__all__ = ["extract_python_items"]
