+++ src/specleft/discovery/miners/typescript/__init__.py
@@ -1,0 +1,8 @@
+# SPDX-License-Identifier: Apache-2.0
+# Copyright (c) 2026 SpecLeft Contributors
+
+"""TypeScript/JavaScript-specific discovery miners."""
+
+from specleft.discovery.miners.typescript.jsdoc import extract_jsdoc_items
+
+__all__ = ["extract_jsdoc_items"]
